JavaScript file update final commit
Added the variable that gives the current date and time. Added the variable that takes the time that the users computer has and put the time on the page. Created formula to decide witch image to display on the web page. Added the variable that Displays the certain sky image. Added comments. Ready for review.